# api-gateway/docs/swagger.yaml
basePath: /
definitions:
  main.AuthRequest:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  main.AuthResponse:
    properties:
      code:
        type: integer
      data:
        properties:
          access_token:
            type: string
          id:
            type: string
        type: object
      status:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: API Gateway для авторизации и управления книгами.
  title: API Gateway
  version: "1.0"
paths:
  /auth_v1/login:
    post:
      consumes:
      - application/json
      description: Авторизует пользователя
      parameters:
      - description: Данные пользователя
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/main.AuthRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.AuthResponse'
        "400":
          description: Invalid request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Авторизация
      tags:
      - Auth
  /auth_v1/register:
    post:
      consumes:
      - application/json
      description: Регистрирует нового пользователя
      parameters:
      - description: Данные пользователя
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/main.AuthRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.AuthResponse'
        "400":
          description: Invalid request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Регистрация
      tags:
      - Auth
  /books:
    get:
      description: Получает список книг из сервиса товаров
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Failed to connect to service
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Получение списка книг
      tags:
      - Books
    post:
      consumes:
      - application/json
      description: Добавляет книгу в сервис товаров
      parameters:
      - description: Данные книги
        in: body
        name: request
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Failed to connect to service
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Добавление книги
      tags:
      - Books
swagger: "2.0"